requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==41.0.7
defusedxml==0.7.1
Django==5.0.1
django-allauth==0.60.1
django-ckeditor==6.7.0
django-crispy-forms==2.1
django-js-asset==2.2.0
Faker==22.2.0
idna==3.6
oauthlib==3.2.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stripe==8.8.0
typing_extensions==4.10.0
urllib3==2.1.0
